+++ server/src/bruFiles/completions/mapVariablesToCompletions.ts
@@ -1,11 +1,6 @@
+import { VariableReferenceType, Range, Position } from "@global_shared";
 import {
-    VariableReferenceType,
-    Range,
-    BrunoVariableReference,
-    Position,
-} from "@global_shared";
-import {
-    EnvVariableCommonRequestData,
+    VariableSpecificRequestData,
     groupReferencesByName,
     mapStaticEnvVariablesToCompletions,
     ReferenceFromOwnFileDetails,
@@ -15,29 +10,24 @@ import {
     CompletionItemKind,
     TextEdit,
 } from "vscode-languageserver";
-import { EquivalentDynamicReferencesFromOtherFiles } from "../shared/interfaces";
+import {
+    EquivalentDynamicReferencesFromOtherFiles,
+    MatchingDynamicVariables,
+} from "../shared/interfaces";
 
-interface MatchingDynamicEnvVariables {
-    fromSameFile: {
-        blockName: string;
-        variableReference: BrunoVariableReference;
-    }[];
-    fromOtherFiles: EquivalentDynamicReferencesFromOtherFiles[];
-}
-
-export function mapEnvVariablesToCompletions(
+export function mapVariablesToCompletions(
     matchingStaticEnvVariables: {
         environmentFile: string;
         matchingVariableKeys: string[];
         isConfiguredEnv: boolean;
     }[],
-    matchingDynamicEnvVariables: MatchingDynamicEnvVariables,
-    requestData: EnvVariableCommonRequestData,
+    matchingDynamicVariables: MatchingDynamicVariables,
+    requestData: VariableSpecificRequestData,
     appendOnInsertion?: string,
 ) {
-    const resultsForDynamicVariables = mapDynamicEnvVariables(
+    const resultsForDynamicVariables = mapDynamicVariables(
         requestData,
-        matchingDynamicEnvVariables,
+        matchingDynamicVariables,
         {
             prefixForSortText: "a",
             appendOnInsertion,
@@ -48,7 +38,7 @@ export function mapEnvVariablesToCompletions(
         requestData,
         filterOutStaticVariablesWithDynamicReferences(
             matchingStaticEnvVariables,
-            matchingDynamicEnvVariables,
+            matchingDynamicVariables,
         ),
         // Display static environment variables below dynamic ones.
         { prefixForSortText: "b", appendOnInsertion },
@@ -57,15 +47,15 @@ export function mapEnvVariablesToCompletions(
     return resultsForDynamicVariables.concat(resultsForStaticVariables);
 }
 
-function mapDynamicEnvVariables(
-    requestData: EnvVariableCommonRequestData,
-    { fromSameFile, fromOtherFiles }: MatchingDynamicEnvVariables,
+function mapDynamicVariables(
+    requestData: VariableSpecificRequestData,
+    { fromSameFile, fromOtherFiles }: MatchingDynamicVariables,
     modifications: {
         prefixForSortText: string;
         appendOnInsertion?: string;
     },
 ) {
-    const groupedRefs = groupReferencesByName(fromSameFile, fromOtherFiles);
+    const groupedRefs = groupReferencesByName({ fromSameFile, fromOtherFiles });
 
     return groupedRefs
         .map((ref) => {
@@ -100,7 +90,7 @@ function mapDynamicEnvVariables(
 }
 
 function getCompletionForRefsWithinOwnFile(
-    { variable: { start, end } }: EnvVariableCommonRequestData,
+    { variable: { start, end } }: VariableSpecificRequestData,
     groupedReferences: {
         variableName: string;
         referenceType: VariableReferenceType;
@@ -162,7 +152,7 @@ function getCompletionForRefsWithinOwnFile(
 }
 
 function getCompletionForRefsFromOnlyOtherFiles(
-    { variable: { start, end } }: EnvVariableCommonRequestData,
+    { variable: { start, end } }: VariableSpecificRequestData,
     groupedReferences: {
         variableName: string;
         referenceType: VariableReferenceType;
@@ -185,7 +175,7 @@ function getCompletionForRefsFromOnlyOtherFiles(
     return {
         label: variableName,
         labelDetails: {
-            description: `${mostRelevantReference.relativePathToSourceFile}`,
+            description: `${mostRelevantReference.path.relativeToSourceFile}`,
         },
         kind: getKind(referenceType),
         detail:
@@ -212,7 +202,7 @@ function filterOutStaticVariablesWithDynamicReferences(
         matchingVariableKeys: string[];
         isConfiguredEnv: boolean;
     }[],
-    matchingDynamicEnvVariables: MatchingDynamicEnvVariables,
+    matchingDynamicEnvVariables: MatchingDynamicVariables,
 ) {
     const allVariableNamesFromDynamicReferences =
         matchingDynamicEnvVariables.fromSameFile