+++ server/src/bruFiles/hover/getHoverForNonCodeBlock.ts
@@ -1,11 +1,13 @@
 import {
     Block,
+    BrunoVariableType,
     getBlocksWithoutVariableSupport,
     getDictionaryBlockArrayField,
     getExistingRequestFileTags,
     getVariableForPositionInNonCodeBlock,
     Logger,
     MetaBlockKey,
+    Range,
     RequestFileBlockName,
     TagOccurences,
     TextDocumentHelper,
@@ -17,7 +19,7 @@ import {
 } from "../../shared";
 import { basename } from "path";
 import { Hover } from "vscode-languageserver";
-import { getHoverForEnvVariable } from "./getHoverForEnvVariable";
+import { getHoverForBrunoVariable } from "./getHoverForBrunoVariable";
 import { BlockRequestWithAdditionalData } from "../shared/interfaces";
 
 export function getHoverForNonCodeBlock(
@@ -25,14 +27,13 @@ export function getHoverForNonCodeBlock(
     params: BlockRequestWithAdditionalData<Block>,
     configuredEnvironmentName?: string,
 ) {
-    return (
-        getHoverForTagsInMetaBlock(itemProvider, params) ??
-        getHoverForVariablesInNonCodeBlocks(
-            params,
-            params.request.documentHelper,
-            configuredEnvironmentName,
-        )
-    );
+    return params.file.blockContainingPosition.name == RequestFileBlockName.Meta
+        ? getHoverForTagsInMetaBlock(itemProvider, params)
+        : getHoverForVariablesInNonCodeBlocks(
+              params,
+              params.request.documentHelper,
+              configuredEnvironmentName,
+          );
 }
 
 function getHoverForTagsInMetaBlock(
@@ -114,11 +115,15 @@ function getHoverForVariablesInNonCodeBlocks(
         return undefined;
     }
 
-    return getHoverForEnvVariable(
+    return getHoverForBrunoVariable(
         fullRequest,
-        variable.name,
-        // In non-code blocks, variables can not be set.
-        VariableReferenceType.Read,
+        {
+            variableName: variable.name,
+            variableNameRange: new Range(variable.start, variable.end),
+            variableType: BrunoVariableType.Unknown,
+            // In non-code blocks, variables can not be set.
+            referenceType: VariableReferenceType.Read,
+        },
         configuredEnvironmentName,
     );
 }